Split Response class into it's own file
